# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FitKart API
  description: |
    FitKart is a mobile fitness rewards platform that tracks user steps and allows redemption for products.
    This is the official RESTful API specification for all client integrations.
  version: 1.0.0
  contact:
    name: FitKart API Support
    email: api-support@fitkart.com
    url: https://fitkart.com/support
  license:
    name: Proprietary
    url: https://fitkart.com/legal/license
servers:
  - url: https://api.fitkart.com/api
    description: Production Server
  - url: https://staging-api.fitkart.com/api
    description: Staging Server
  - url: http://localhost:3000/api
    description: Development Server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - full_name
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
                full_name:
                  type: string
                  example: John Doe
                phone:
                  type: string
                  example: "+14155552671"
                country_code:
                  type: string
                  example: "+1"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/otp/send:
    post:
      tags:
        - Authentication
      summary: Request OTP code
      operationId: sendOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  message:
                    type: string
                    example: "OTP sent to john@example.com"
  /auth/otp/verify:
    post:
      tags:
        - Authentication
      summary: Verify OTP code
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp_code
              properties:
                email:
                  type: string
                  format: email
                otp_code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - error
      properties:
        status:
          type: string
          enum: [error]
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "AUTH_INVALID_CREDENTIALS"
            message:
              type: string
              example: "Invalid email or password"
            details:
              type: object
              additionalProperties: true
    AuthResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          required:
            - id
            - email
            - access_token
            - expires_in
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            full_name:
              type: string
            access_token:
              type: string
            refresh_token:
              type: string
            expires_in:
              type: integer
              description: Seconds until token expiry
            token_type:
              type: string
              enum: [Bearer]
    TokenResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            access_token:
              type: string
            expires_in:
              type: integer
...